+++ package.json
@@ -2,10 +2,14 @@
   "name": "psychostudios_redesign",
   "version": "1.0.0",
   "description": "Psycho Studios rebranded web app",
-  "main": "index.js",
+  "main": "server.js",
   "type": "module",
   "scripts": {
-    "compile": "sass --watch --no-source-map ./public/scss/main.scss public/css/main.css"
+    "start": "nodemon server.js",
+    "compile": "sass --watch --no-source-map ./public/scss/main.scss public/css/main.css",
+    "watch-app": "index.js",
+    "watch-css": "sass --watch --no-source-map ./public/scss/main.scss public/css/main.css",
+    "watch": "npm-run-all --parallel"
   },
   "repository": {
     "type": "git",
@@ -19,6 +23,10 @@
   "homepage": "https://github.com/Jeshawn/PsychoStudios_Redesign/tree/development#readme",
   "dependencies": {
     "express": "^4.18.2",
+    "jquery": "^3.7.1",
     "sass": "^1.64.1"
+  },
+  "devDependencies": {
+    "npm-run-all": "^4.1.5"
   }
 }
