+++ public/scss/layout/_about.scss
@@ -3,3 +3,7 @@
 .about-section { 
     background-color: #57a7f1;
 }
+
+div.row {
+    margin: 3rem;
+}
