+++ public/scss/layout/_masthead.scss
@@ -5,7 +5,7 @@
   padding-top: calc(6rem + 74px);
   padding-bottom: 6rem;
   .masthead-heading {
-    font-size: 2.75rem;
+    font-size: 5rem;
     line-height: 2.75rem;
   }
   .masthead-subheading {
