+++ public/success.html
@@ -1,0 +1,140 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <meta name="description" content="">
+  <meta name="author" content="">
+
+  <title>Psycho Studios</title>
+  
+  <link rel="icon" href="/public/images/newpsychostudios.png" type="image/x-icon">
+
+  <!-- Custom fonts for this theme -->
+  <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
+  <link href="https://fonts.googleapis.com/css?family=Montserrat:400,700" rel="stylesheet" type="text/css">
+  <link href="https://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic" rel="stylesheet" type="text/css">
+
+  <!-- Theme CSS -->
+  <link href="css/main.css" rel="stylesheet">
+    <!-- Bootstrap cdn -->
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" 
+        rel="stylesheet" 
+        integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" 
+        crossorigin="anonymous">
+  <!-- font awesome cdn -->
+  <script defer src="https://kit.fontawesome.com/f030b35153.js" crossorigin="anonymous"></script>
+   <!-- Bootstrap JS bundle cdn -->
+  <script defer src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" 
+          integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" 
+          crossorigin="anonymous">
+  </script>
+
+</head>
+
+<body id="page-top">
+
+  <!-- Navigation -->
+  <nav class="navbar navbar-expand-lg sticky-top" id="mainNav">
+    <div class="container-fluid">
+      <a class="navbar-brand" href="index.html">PsychoStudios</a>
+      <button class="navbar-toggler bg-warning" 
+              type="button" 
+              data-bs-toggle="collapse" 
+              data-bs-target="#navBar" 
+              aria-controls="navBar" 
+              aria-expanded="false" 
+              aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      
+      <div class="collapse navbar-collapse" id="navBar">
+        <ul class="navbar-nav">
+          <li class="nav-item mb-2 px-3">
+            <a class="nav-link js-scroll-trigger" href="#portfolio">Current Projects</a>
+          </li>
+          <li class="nav-item px-3">
+            <a class="nav-link  js-scroll-trigger" href="#about">About</a>
+          </li>
+          <li class="nav-item px-3" >
+            <a class="nav-link  js-scroll-trigger" href="#contact">Contact</a>
+          </li>
+        </ul>
+      </div>
+    </div>
+  </nav>
+
+
+
+
+
+<!-- Footer -->
+<footer class="footer text-center">
+    <div class="container">
+      <div class="row">
+
+        <!-- Footer Location -->
+        <div class="col-lg-4 mb-5 mb-lg-0">
+          <h4 class="text-black  mb-4">Location</h4>
+          <p class="lead mb-0"> Metro Detrot Area
+            <br>Michigan</p>
+        </div>
+
+        <!-- Footer Social Icons -->
+        <div class="col-lg-4 mb-5 mb-lg-0">
+          <h4 class="text-uppercase mb-4">Connect with Us</h4>
+          <a class="btn btn-outline-dark btn-social mx-1" 
+             href="https://twitter.com/JChrite" 
+             target="_blank">
+            <i class="fab fa-fw fa-twitter">
+            </i>
+          </a>
+          <a class="btn btn-outline-dark btn-social mx-1" 
+             href="https://www.linkedin.com/in/jeshawn-chrite"  
+             target="_blank">
+            <i class="fab fa-fw fa-linkedin-in">
+            </i>
+          </a>
+          <a class="btn btn-outline-dark  btn-social mx-1" 
+             href="https://instagram.com/" 
+             target="_blank">
+            <i class="fa-brands fa-instagram fa-xl">
+            </i>
+          </a>
+          <a class="btn btn-outline-dark btn-social mx-1" 
+             href="https://youtube.com/"
+             target="_blank">
+            <i class="fa-brands fa-youtube fa-xl">
+            </i>
+          </a>
+        </div>
+      </div>
+    </div>
+  </footer>
+
+  <!-- Copyright Section -->
+  <section class="copyright py-4 text-center text-black">
+    <div class="container">
+      <small> created by psychostudios <span>&copy;</span></small>
+    </div>
+  </section>
+
+  <!-- Scroll to Top Button (Only visible on small and extra-small screen sizes) -->
+  <div class="scroll-to-top d-lg-none position-fixed ">
+    <a class="js-scroll-trigger d-block text-center text-white rounded" href="#page-top">
+      <i class="fa fa-chevron-up"></i>
+    </a>
+  </div>
+
+
+
+
+
+  
+<script src="/js/index.js"></script>
+
+</body>
+
+</html>
